chore: minimize loader diagnostics changes
